Lógica de validação para regras de negócio e Concluindo e testando as regras de negócio de autores

+++ src/main/java/io/github/andreynicollas/api_libary/service/AutorService.java
@@ -1,7 +1,10 @@
 package io.github.andreynicollas.api_libary.service;
 
+import io.github.andreynicollas.api_libary.exceptions.OperacaoNaoPermitidaException;
 import io.github.andreynicollas.api_libary.model.Autor;
 import io.github.andreynicollas.api_libary.repository.AutorRepository;
+import io.github.andreynicollas.api_libary.repository.LivroRepository;
+import io.github.andreynicollas.api_libary.validator.AutorValidator;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -12,12 +15,17 @@ import java.util.UUID;
 public class AutorService {
 
     private final AutorRepository repository;
+    private final AutorValidator validator;
+    private final LivroRepository livroRepository;
 
-    public AutorService(AutorRepository repository) {
+    public AutorService(AutorRepository repository, AutorValidator validator, LivroRepository livroRepository) {
         this.repository = repository;
+        this.validator = validator;
+        this.livroRepository = livroRepository;
     }
 
     public Autor salvarAutor(Autor autor) {
+        validator.validar(autor);
         return repository.save(autor);
     }
 
@@ -25,6 +33,7 @@ public class AutorService {
         if (autor.getId() == null) {
             throw new IllegalArgumentException("Para atualizar, é necessário que o autor já esteja salvo na base!");
         }
+        validator.validar(autor);
         repository.save(autor);
     }
 
@@ -33,6 +42,9 @@ public class AutorService {
     }
 
     public void deletar(Autor autor) {
+        if (possuiLivro(autor)) {
+            throw new OperacaoNaoPermitidaException("Não é permitido excluir um Autor que possui livros cadastrados!");
+        }
         repository.delete(autor);
     }
 
@@ -50,5 +62,9 @@ public class AutorService {
         }
 
         return repository.findAll();
+    }
+
+    public boolean possuiLivro(Autor autor) {
+        return livroRepository.existsByAutor(autor);
     }
 }
